Add DAG run metadata to lakeFS merge commits
Fixes #57.
